+++ src/processes/TypeAssure.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace koolreport\processes;
+
+class TypeAssure extends \koolreport\core\Process
+{
+    protected function onInput($row)
+    {
+        foreach($row as $k=>$v) {
+            if(gettype($v)==="string" && $this->metaData["columns"][$k]["type"]==="number") {
+                $row[$k] = (float)$v;
+            }
+        }
+        $this->next($row);
+    }
+}