+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 2. `Count`: Adding new Count process to count rows with condition.
 3. `DataStore`: Add method `getScalar()`
 4. `Filter`: Adding operator `<>`
+5. `TypeAssure`: New type conversion process
 
 ## Version 4.7.1